Move PollFileChangesLoop() to LiveTextFile.
